requirements.txt
python-dotenv
requests
sense-hat